split into AbstractThermalStorageInput
