+++ src/main/java/edu/ie3/datamodel/models/input/thermal/DomesticHotWaterStorageInput.java
@@ -14,7 +14,7 @@ import javax.measure.quantity.Temperature;
 import javax.measure.quantity.Volume;
 import tech.units.indriya.ComparableQuantity;
 
-public class DomesticHotWaterStorageInput extends CylindricalStorageInput {
+public class DomesticHotWaterStorageInput extends AbstractThermalStorageInput {
 
   /**
    * Constructor for DomesticHotWaterStorageInput
@@ -78,39 +78,8 @@ public class DomesticHotWaterStorageInput extends CylindricalStorageInput {
     super(uuid, id, bus, storageVolumeLvl, inletTemp, returnTemp, c, pThermalMax);
   }
 
-  @Override
-  public DomesticHotWaterStorageInputCopyBuilder copy() {
-    return new DomesticHotWaterStorageInputCopyBuilder(this);
-  }
-
-  @Override
-  public String toString() {
-    return "DomesticHotWaterStorageInput{"
-        + "uuid="
-        + getUuid()
-        + ", id="
-        + getId()
-        + ", operator="
-        + getOperator().getUuid()
-        + ", operationTime="
-        + getOperationTime()
-        + ", bus="
-        + getThermalBus().getUuid()
-        + ", storageVolumeLvl="
-        + getStorageVolumeLvl()
-        + ", inletTemp="
-        + getInletTemp()
-        + ", returnTemp="
-        + getReturnTemp()
-        + ", c="
-        + getC()
-        + ", pThermalMax="
-        + getpThermalMax()
-        + '}';
-  }
-
   public static class DomesticHotWaterStorageInputCopyBuilder
-      extends CylindricalStorageInputCopyBuilder {
+      extends AbstractThermalStorageInputCopyBuilder {
 
     protected DomesticHotWaterStorageInputCopyBuilder(DomesticHotWaterStorageInput entity) {
       super(entity);
